+++ forge/src/main/java/net/xstopho/resource_backpacks/datagen/BlockTagProv.java
@@ -1,0 +1,21 @@
+package net.xstopho.resource_backpacks.datagen;
+
+import net.minecraft.core.HolderLookup;
+import net.minecraft.data.PackOutput;
+import net.minecraftforge.common.data.BlockTagsProvider;
+import net.minecraftforge.common.data.ExistingFileHelper;
+import net.xstopho.resource_backpacks.BackpackConstants;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.concurrent.CompletableFuture;
+
+public class BlockTagProv extends BlockTagsProvider {
+    public BlockTagProv(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider, @Nullable ExistingFileHelper existingFileHelper) {
+        super(output, lookupProvider, BackpackConstants.MOD_ID, existingFileHelper);
+    }
+
+    @Override
+    protected void addTags(HolderLookup.Provider provider) {
+
+    }
+}